refactor: move Footer & Authors to a folders, move styles in separate files
